Ongoing libs3 and s3/lambda.py and services.py refactor to reuse methods on report, please @victorskl #reviewthis carefully
