+++ data_processors/s3/services.py
@@ -1,51 +1,16 @@
 import logging
-import json
-from datetime import datetime
-from typing import Tuple
 
-from django.core.exceptions import ObjectDoesNotExist
 from django.db import transaction
 from django.db.models import ExpressionWrapper, Value, CharField, Q, F
 
-from data_portal.models import S3Object, LIMSRow, S3LIMS, Report
-from utils import libs3
+from data_portal.models import LIMSRow, S3LIMS
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 
-@transaction.atomic
-def persist_s3_object(bucket: str, key: str, last_modified_date: datetime, size: int, e_tag: str) -> Tuple[int, int]:
-    """
-    Persist an s3 object record into the db
-    :param bucket: s3 bucket name
-    :param key: s3 object key
-    :param last_modified_date: s3 object last modified date
-    :param size: s3 object size
-    :param e_tag: s3 objec etag
-    :return: number of s3 object created, number of s3-lims association records created
-    """
-    query_set = S3Object.objects.filter(bucket=bucket, key=key)
-    new = not query_set.exists()
-
-    if new:
-        logger.info(f"Creating a new S3Object (bucket={bucket}, key={key})")
-        s3_object = S3Object(
-            bucket=bucket,
-            key=key
-        )
-    else:
-        logger.info(f"Updating a existing S3Object (bucket={bucket}, key={key})")
-        s3_object: S3Object = query_set.get()
-
-    s3_object.last_modified_date = last_modified_date
-    s3_object.size = size
-    s3_object.e_tag = e_tag
-    s3_object.save()
-
-    if not new:
-        return 0, 0
-
+@transaction.atomic()
+def associate_lims_rows_with_s3_obj(bucket: str, key: str, s3_object: str):
     # TODO remove association logic and drop S3LIMS table, related with global search overhaul
     # Number of s3-lims association records we have created in this run
     new_association_count = 0
@@ -78,69 +43,3 @@ def persist_s3_object(bucket: str, key: str, last_modified_date: datetime, size:
         logger.debug(f"No association to any LIMS row is found for the S3Object (bucket={bucket}, key={key})")
 
     return 1, new_association_count
-
-
-@transaction.atomic
-def delete_s3_object(bucket_name: str, key: str) -> Tuple[int, int]:
-    """
-    Delete a S3 object record from db
-    :param bucket_name: s3 bucket name
-    :param key: s3 object key
-    :return: number of s3 records deleted, number of s3-lims association records deleted
-    """
-    try:
-        s3_object: S3Object = S3Object.objects.get(bucket=bucket_name, key=key)
-        s3_lims_records = S3LIMS.objects.filter(s3_object=s3_object)
-        s3_lims_count = s3_lims_records.count()
-        s3_lims_records.delete()
-        s3_object.delete()
-        logger.info(f"Deleted S3Object: s3://{bucket_name}/{key}")
-        return 1, s3_lims_count
-    except ObjectDoesNotExist as e:
-        logger.info(f"No deletion required. Non-existent S3Object (bucket={bucket_name}, key={key}): {str(e)}")
-        return 0, 0
-
-
-def tag_s3_object(bucket_name: str, key: str):
-    """
-    Tag S3 Object if extension is .bam
-
-    NOTE: You can associate up to 10 tags with an object. See
-    https://docs.aws.amazon.com/AmazonS3/latest/dev/object-tagging.html
-    :param bucket_name:
-    :param key:
-    """
-
-    if key.endswith('.bam'):
-        response = libs3.get_s3_object_tagging(bucket=bucket_name, key=key)
-        tag_set = response.get('TagSet', [])
-
-        tag_archive = {'Key': 'Archive', 'Value': 'true'}
-        tag_bam = {'Key': 'Filetype', 'Value': 'bam'}
-
-        if len(tag_set) == 0:
-            tag_set.append(tag_archive)
-            tag_set.append(tag_bam)
-        else:
-            # have existing tags
-            immutable_tags = tuple(tag_set)  # have immutable copy first
-            if tag_bam not in immutable_tags:
-                tag_set.append(tag_bam)  # just add tag_bam
-            if tag_archive not in immutable_tags:
-                values = set()
-                for tag in immutable_tags:
-                    for value in tag.values():
-                        values.add(value)
-                if tag_archive['Key'] not in values:
-                    tag_set.append(tag_archive)  # only add if Archive is not present
-
-        payload = libs3.put_s3_object_tagging(bucket=bucket_name, key=key, tagging={'TagSet': tag_set})
-
-        if payload['ResponseMetadata']['HTTPStatusCode'] == 200:
-            logger.info(f"Tagged the S3Object ({key}) with ({str(tag_set)})")
-        else:
-            logger.error(f"Failed to tag the S3Object ({key}) with ({str(payload)})")
-
-    else:
-        # sound of silence
-        pass
